+++ src/main/java/tstProj/Data/Interfaces/BooksTypesContainer.java
@@ -6,7 +6,7 @@ import tstProj.Data.Entities.BooksTypes;
 
 public interface BooksTypesContainer {
 	
-	public BooksTypes GetById(long ID);
+	public BooksTypes GetById(Integer ID);
 	
 	public BooksTypes GetByBookType (String t);
